+++ src/models/Encounters/Encounter.js
