+++ src/models/Language/linkingVerbs.js
@@ -1,4 +1,4 @@
-import { capitalizeFirst } from '../utils';
+import { capitalizeFirst } from '../../utils/utils';
 
 
 export const were = actor => (actor.grammar.isPlural
