+++ src/models/Sprite/name/badList.js
