+++ src/models/Sprite/name/soulName.js
@@ -1,7 +1,7 @@
 import { prefixTrie, syllableTrie } from './nameTries';
 import Trie from './Trie';
 import { badList } from './badList';
-import { randomSeed } from '../../../utils';
+import { randomSeed } from '../../../utils/utils';
 
 const MIN_LENGTH = 5;
