+++ src/models/Sprite/spriteEncyclopedia.js
