+++ src/models/Sprite/spriteTrust.js
