+++ src/models/Sprite/templates/singTemplates.js
@@ -1,7 +1,7 @@
-import { They_re } from '../pronouns';
+import { They_re } from '../../Language/pronouns';
 import { chirp, chirling, a_chirling, Kimberly, chirping,
-  purrs, chirps, growls } from '../../modules/Sprite/interactions';
-import { spriteText } from '../../modules/Sprite/spriteHelpers';
+  purrs, chirps, growls } from '../interactions';
+import { spriteText } from '../spriteHelpers';
 
 const oldMacDonaldSong = sprite => spriteText(sprite)`
   Old MacDonald had a den${'  '}
