+++ src/sharedComponents/SpriteHeadshot/SpriteHeadshot.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import './SpriteHeadshot.css';
-import { SPRITE_ENCYCLOPEDIA } from '../../modules/Sprite/spriteEncyclopedia';
+import { SPRITE_ENCYCLOPEDIA } from '../../models/Sprite/spriteEncyclopedia';
 
 const SPRITE_SIZE = 350;
